fullstack/backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
boto3==1.38.0
botocore==1.38.0
cachetools==5.5.1
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
cssselect==1.3.0
cssutils==2.11.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emails==0.6
Faker==37.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.18.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.7
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.5
jiter==0.9.0
jmespath==1.0.1
lxml==5.4.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mypy==1.15.0
mypy-extensions==1.1.0
nodeenv==1.9.1
nulltype==2.3.1
oauthlib==3.2.2
openai==1.75.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
plaid-python==30.0.0
platformdirs==4.3.6
pluggy==1.5.0
premailer==3.10.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.1
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.7.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
ruff==0.11.6
s3transfer==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
starlette==0.45.3
stripe==12.0.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.15.1
types-passlib==1.7.7.20241221
types-requests==2.32.0.20241016
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.4
websockets==15.0.1
